wrapping CedarCopilot in server component, bumping package, exporting tiptap components
